FIX Add entrypoint to `runtime` images
This way the notebook server starts automatically and the images support passing commands like `python -c 'import cuml'`. Without the entrypoint definition the `rapids` conda environment is not activated and we only see the base environment. The `.run_in_rapids` file is overridden with a runtime version in the notebooks install partial template.