FIX Correct Dockerfile selection in gpuCI script
The previous selector chose the `centos7` Dockerfiles for `ubuntu16.04` and `ubuntu18.04`. This resulted in errors with the build that prevented the ENTRYPOINT to work correctly and despite local builds working, the produced builds from gpuCI did not have the correct `/rapids` directory and the necessary scripts/notebooks as expected.

I tested these changes in the gpuCI script job and confirmed they work as expected.