Remove `libm.so.6`
This PR removes `libm.so.6` since it was added a long time ago and it's purpose is no longer clear. Additionally, there are some `21.12` build errors related to this file in the latest nightlies. If the images build and run successfully after this file is removed, then no further action is necessary.