Remove `~/.m2` directory after `dask-sql` source build
This PR removes the `~/.m2` directory from the images after the `dask-sql` source build. This change is necessary because the `~/.m2` directory contains `log4j`, which triggers some security alerts for our images. As per @jdye64, the `log4j` dependency is only used during the compile steps (and is not in the resulting `DaskSQL.jar` file) and is therefore safe to simply remove with `rm -rf ~/.m2`.